Change schema of http_sessions
Looking through the access patterns in this code, I can see that the id
column is not used anywhere and lookups are done on un-indexed columns.
This change remedies some things to improve scalability and to adhere
to some design best practices.

Replace the old primary key column (id) with an opaque identifier which
is already present on the PGSession struct. It should be difficult to
guess another identifier that could be looked up. This is a protection
against session fixation attacks, where a session identifier could be
modified by the client and sent to the server in order to impersonate
another user. This wouldn't be a problem if the server does an
authentication check on the session (already provided by the gorilla
sessions library), but I think this extra defense in depth is worth the
effort to handle cases where the server may not authenticate sessions
for whatever reason.

Help the Cleanup process lookup expired sessions more quickly by adding
an index to expires_on.